+++ src/bin/conserve.rs
@@ -28,6 +28,7 @@ use time::UtcOffset;
 #[allow(unused_imports)]
 use tracing::{debug, error, info, trace, warn, Level};
 
+use crate::transport::Transport;
 use conserve::termui::{enable_tracing, TermUiMonitor, TraceTimeStyle};
 use conserve::*;
 
@@ -320,7 +321,7 @@ impl std::process::Termination for ExitCode {
 
 impl Command {
     fn run(&self, monitor: Arc<TermUiMonitor>) -> Result<ExitCode> {
-        let mut stdout = std::io::stdout();
+        let mut stdout = io::stdout();
         match self {
             Command::Backup {
                 archive,
@@ -342,7 +343,7 @@ impl Command {
                     ..Default::default()
                 };
                 let stats = backup(
-                    &Archive::open(open_transport(archive)?)?,
+                    &Archive::open(Transport::new(archive)?)?,
                     source,
                     &options,
                     monitor,
@@ -353,7 +354,7 @@ impl Command {
             }
             Command::Debug(Debug::Blocks { archive }) => {
                 let mut bw = BufWriter::new(stdout);
-                for hash in Archive::open(open_transport(archive)?)?
+                for hash in Archive::open(Transport::new(archive)?)?
                     .block_dir()
                     .blocks(monitor)?
                     .collect::<Vec<BlockHash>>()
@@ -368,7 +369,7 @@ impl Command {
             }
             Command::Debug(Debug::Referenced { archive }) => {
                 let mut bw = BufWriter::new(stdout);
-                let archive = Archive::open(open_transport(archive)?)?;
+                let archive = Archive::open(Transport::new(archive)?)?;
                 for hash in archive.referenced_blocks(&archive.list_band_ids()?, monitor)? {
                     writeln!(bw, "{hash}")?;
                 }
@@ -376,7 +377,7 @@ impl Command {
             Command::Debug(Debug::Unreferenced { archive }) => {
                 print!(
                     "{}",
-                    Archive::open(open_transport(archive)?)?
+                    Archive::open(Transport::new(archive)?)?
                         .unreferenced_blocks(monitor)?
                         .map(|hash| format!("{}\n", hash))
                         .collect::<Vec<String>>()
@@ -390,7 +391,7 @@ impl Command {
                 break_lock,
                 no_stats,
             } => {
-                let stats = Archive::open(open_transport(archive)?)?.delete_bands(
+                let stats = Archive::open(Transport::new(archive)?)?.delete_bands(
                     backup,
                     &DeleteOptions {
                         dry_run: *dry_run,
@@ -433,7 +434,7 @@ impl Command {
                 break_lock,
                 no_stats,
             } => {
-                let archive = Archive::open(open_transport(archive)?)?;
+                let archive = Archive::open(Transport::new(archive)?)?;
                 let stats = archive.delete_bands(
                     &[],
                     &DeleteOptions {
@@ -447,7 +448,7 @@ impl Command {
                 }
             }
             Command::Init { archive } => {
-                Archive::create(open_transport(archive)?)?;
+                Archive::create(Transport::new(archive)?)?;
                 debug!("Created new archive in {archive:?}");
             }
             Command::Ls {
@@ -532,7 +533,7 @@ impl Command {
                 no_stats,
             } => {
                 let band_selection = band_selection_policy_from_opt(backup);
-                let archive = Archive::open(open_transport(archive)?)?;
+                let archive = Archive::open(Transport::new(archive)?)?;
                 let _ = no_stats; // accepted but ignored; we never currently print stats
                 let options = RestoreOptions {
                     exclude: Exclude::from_patterns_and_files(exclude, exclude_from)?,
@@ -575,7 +576,7 @@ impl Command {
                 let options = ValidateOptions {
                     skip_block_hashes: *quick,
                 };
-                Archive::open(open_transport(archive)?)?.validate(&options, monitor.clone())?;
+                Archive::open(Transport::new(archive)?)?.validate(&options, monitor.clone())?;
                 if monitor.error_count() != 0 {
                     warn!("Archive has some problems.");
                 } else {
@@ -594,7 +595,7 @@ impl Command {
                 } else {
                     Some(*LOCAL_OFFSET.read().unwrap())
                 };
-                let archive = Archive::open(open_transport(archive)?)?;
+                let archive = Archive::open(Transport::new(archive)?)?;
                 let options = ShowVersionsOptions {
                     newest_first: *newest,
                     tree_size: *sizes,
@@ -610,7 +611,7 @@ impl Command {
 }
 
 fn stored_tree_from_opt(archive_location: &str, backup: &Option<BandId>) -> Result<StoredTree> {
-    let archive = Archive::open(open_transport(archive_location)?)?;
+    let archive = Archive::open(Transport::new(archive_location)?)?;
     let policy = band_selection_policy_from_opt(backup);
     archive.open_stored_tree(policy)
 }