+++ src/index.rs
@@ -19,6 +19,7 @@ use std::path::Path;
 use std::sync::Arc;
 use std::vec;
 
+use crate::transport::Transport;
 use itertools::Itertools;
 use time::OffsetDateTime;
 use tracing::{debug, debug_span, error};
@@ -28,7 +29,6 @@ use crate::counters::Counter;
 use crate::entry::KindMeta;
 use crate::monitor::Monitor;
 use crate::stats::IndexReadStats;
-use crate::transport::local::LocalTransport;
 use crate::unix_time::FromUnixAndNanos;
 use crate::*;
 
@@ -177,7 +177,7 @@ impl IndexEntry {
 /// hunks preserve apath order.
 pub struct IndexWriter {
     /// The `i` directory within the band where all files for this index are written.
-    transport: Arc<dyn Transport>,
+    transport: Transport,
 
     /// Currently queued entries to be written out, in arbitrary order.
     entries: Vec<IndexEntry>,
@@ -199,7 +199,7 @@ pub struct IndexWriter {
 /// Accumulate and write out index entries into files in an index directory.
 impl IndexWriter {
     /// Make a new builder that will write files into the given directory.
-    pub fn new(transport: Arc<dyn Transport>) -> IndexWriter {
+    pub fn new(transport: Transport) -> IndexWriter {
         IndexWriter {
             transport,
             entries: Vec::new(),
@@ -278,7 +278,7 @@ fn hunk_relpath(hunk_number: u32) -> String {
 /// Utility to read the stored index
 pub struct IndexRead {
     /// Transport pointing to this index directory.
-    transport: Arc<dyn Transport>,
+    transport: Transport,
 
     /// Decompressor for the index to read
     decompressor: Decompressor,
@@ -289,11 +289,12 @@ pub struct IndexRead {
 
 impl IndexRead {
     #[allow(unused)]
+    // TODO: Deprecate, use Transport?
     pub(crate) fn open_path(path: &Path) -> IndexRead {
-        IndexRead::open(Arc::new(LocalTransport::new(path)))
+        IndexRead::open(Transport::local(path))
     }
 
-    pub(crate) fn open(transport: Arc<dyn Transport>) -> IndexRead {
+    pub(crate) fn open(transport: Transport) -> IndexRead {
         IndexRead {
             transport,
             decompressor: Decompressor::new(),
@@ -367,11 +368,9 @@ impl IndexRead {
         let _span = debug_span!("iter_hunks", ?self.transport).entered();
         let hunks = self.hunks_available().expect("hunks available"); // TODO: Don't panic
         debug!(?hunks);
-
         IndexHunkIter {
             hunks: hunks.into_iter(),
             index: self,
-
             after: None,
         }
     }
@@ -382,7 +381,6 @@ impl IndexRead {
         IndexHunkIter {
             hunks,
             index: self,
-
             after: None,
         }
     }
@@ -394,7 +392,6 @@ impl IndexRead {
 pub struct IndexHunkIter {
     hunks: std::vec::IntoIter<u32>,
     pub index: IndexRead,
-
     /// If set, yield only entries ordered after this apath.
     after: Option<Apath>,
 }
@@ -554,7 +551,7 @@ mod tests {
 
     fn setup() -> (TempDir, IndexWriter) {
         let testdir = TempDir::new().unwrap();
-        let ib = IndexWriter::new(Arc::new(LocalTransport::new(testdir.path())));
+        let ib = IndexWriter::new(Transport::local(testdir.path()));
         (testdir, ib)
     }