+++ src/transport.rs
@@ -1,4 +1,4 @@
-// Copyright 2020, 2021, 2022, 2023 Martin Pool.
+// Copyright 2020-2024 Martin Pool.
 
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,7 +14,7 @@
 //!
 //! Transport operations return std::io::Result to reflect their narrower focus.
 
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::{error, fmt, io, result};
 
@@ -25,38 +25,11 @@ use url::Url;
 use crate::*;
 
 pub mod local;
-use local::LocalTransport;
+#[cfg(feature = "sftp")]
+pub mod sftp;
 
 #[cfg(feature = "s3")]
 pub mod s3;
-
-/// Open a `Transport` to access a local directory.
-///
-/// `s` may be a local path or a URL.
-pub fn open_transport(s: &str) -> crate::Result<Arc<dyn Transport>> {
-    if let Ok(url) = Url::parse(s) {
-        match url.scheme() {
-            "file" => Ok(Arc::new(LocalTransport::new(
-                &url.to_file_path().expect("extract URL file path"),
-            ))),
-            #[cfg(feature = "s3")]
-            "s3" => Ok(s3::S3Transport::new(&url)?),
-            d if d.len() == 1 => {
-                // Probably a Windows path with drive letter, like "c:/thing", not actually a URL.
-                Ok(Arc::new(LocalTransport::new(Path::new(s))))
-            }
-            other => Err(crate::Error::UrlScheme {
-                scheme: other.to_owned(),
-            }),
-        }
-    } else {
-        Ok(Arc::new(LocalTransport::new(Path::new(s))))
-    }
-}
-
-pub fn open_local_transport(path: &Path) -> crate::Result<Arc<dyn Transport>> {
-    Ok(Arc::new(LocalTransport::new(path)))
-}
 
 /// Abstracted filesystem IO to access an archive.
 ///
@@ -66,26 +39,107 @@ pub fn open_local_transport(path: &Path) -> crate::Result<Arc<dyn Transport>> {
 /// A transport has a root location, which will typically be the top directory of the Archive.
 /// Below that point everything is accessed with a relative path, expressed as a PathBuf.
 ///
-/// All Transports must be `Send + Sync`, so they can be passed across or shared across threads.
+/// Transport objects can be cheaply cloned.
 ///
 /// Files in Conserve archives have bounded size and fit in memory so this does not need to
 /// support streaming or partial reads and writes.
-pub trait Transport: Send + Sync + std::fmt::Debug {
-    /// List a directory, separating out file and subdirectory names.
-    ///
-    /// Names are in the arbitrary order that they're returned from the transport.
-    ///
-    /// Any error during iteration causes overall failure.
-    fn list_dir(&self, relpath: &str) -> Result<ListDir>;
+#[derive(Clone)]
+pub struct Transport {
+    protocol: Arc<dyn Protocol + 'static>,
+}
+
+impl Transport {
+    /// Open a new local transport addressing a filesystem directory.
+    pub fn local(path: &Path) -> Self {
+        Transport {
+            protocol: Arc::new(local::Protocol::new(path)),
+        }
+    }
+
+    /// Open a new transport from a string that might be a URL or local path.
+    pub fn new(s: &str) -> Result<Self> {
+        if let Ok(url) = Url::parse(s) {
+            Transport::from_url(&url)
+        } else {
+            Ok(Transport::local(Path::new(s)))
+        }
+    }
+
+    pub fn from_url(url: &Url) -> Result<Self> {
+        let protocol: Arc<dyn Protocol> = match url.scheme() {
+            "file" => Arc::new(local::Protocol::new(
+                &url.to_file_path().expect("extract URL file path"),
+            )),
+            d if d.len() == 1 => {
+                // Probably a Windows path with drive letter, like "c:/thing", not actually a URL.
+                Arc::new(local::Protocol::new(Path::new(url.as_str())))
+            }
+
+            #[cfg(feature = "s3")]
+            "s3" => Arc::new(s3::Protocol::new(url)?),
+
+            #[cfg(feature = "sftp")]
+            "sftp" => Arc::new(sftp::Protocol::new(url)?),
+
+            _other => {
+                return Err(Error {
+                    kind: ErrorKind::UrlScheme,
+                    path: Some(url.as_str().to_owned()),
+                    source: None,
+                })
+            }
+        };
+        Ok(Transport { protocol })
+    }
 
     /// Get one complete file into a caller-provided buffer.
     ///
     /// Files in the archive are of bounded size, so it's OK to always read them entirely into
     /// memory, and this is simple to support on all implementations.
-    fn read_file(&self, path: &str) -> Result<Bytes>;
+    pub fn read_file(&self, path: &str) -> Result<Bytes> {
+        self.protocol.read_file(path)
+    }
+
+    /// List a directory, separating out file and subdirectory names.
+    ///
+    /// Names are in the arbitrary order that they're returned from the transport.
+    ///
+    /// Any error during iteration causes overall failure.
+    pub fn list_dir(&self, relpath: &str) -> Result<ListDir> {
+        self.protocol.list_dir(relpath)
+    }
+
+    /// Make a new transport addressing a subdirectory.
+    pub fn chdir(&self, relpath: &str) -> Self {
+        Transport {
+            protocol: self.protocol.chdir(relpath),
+        }
+    }
+
+    pub fn write_file(&self, relpath: &str, content: &[u8]) -> Result<()> {
+        self.protocol.write_file(relpath, content)
+    }
+
+    pub fn create_dir(&self, relpath: &str) -> Result<()> {
+        self.protocol.create_dir(relpath)
+    }
+
+    pub fn metadata(&self, relpath: &str) -> Result<Metadata> {
+        self.protocol.metadata(relpath)
+    }
+
+    /// Delete a file.
+    pub fn remove_file(&self, relpath: &str) -> Result<()> {
+        self.protocol.remove_file(relpath)
+    }
+
+    /// Delete a directory and all its contents.
+    pub fn remove_dir_all(&self, relpath: &str) -> Result<()> {
+        self.protocol.remove_dir_all(relpath)
+    }
 
     /// Check if a regular file exists.
-    fn is_file(&self, path: &str) -> Result<bool> {
+    pub fn is_file(&self, path: &str) -> Result<bool> {
         match self.metadata(path) {
             Ok(metadata) => Ok(metadata.kind == Kind::File),
             Err(err) if err.kind() == ErrorKind::NotFound => Ok(false),
@@ -93,25 +147,27 @@ pub trait Transport: Send + Sync + std::fmt::Debug {
         }
     }
 
-    /// Create a directory, if it does not exist.
-    ///
-    /// If the directory already exists, it's not an error.
-    ///
-    /// This function does not create missing parent directories.
-    fn create_dir(&self, relpath: &str) -> Result<()>;
+    pub fn url(&self) -> &Url {
+        self.protocol.url()
+    }
 
-    /// Write a complete file.
-    ///
-    /// As much as possible, the file should be written atomically so that it is only visible with
-    /// the complete content. On a local filesystem the content is written to a temporary file and
-    /// then renamed.
-    /// If a temporary file is used, the name should start with `crate::TMP_PREFIX`.
-    ///
-    /// If the file exists it is replaced. (Across transports, and particularly on S3,
-    /// we can't rely on detecting existing files.)
+    #[allow(unused)]
+    fn local_path(&self) -> Option<PathBuf> {
+        self.protocol.local_path()
+    }
+}
+
+impl fmt::Debug for Transport {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "Transport2({})", self.url())
+    }
+}
+
+trait Protocol: Send + Sync {
+    fn read_file(&self, path: &str) -> Result<Bytes>;
     fn write_file(&self, relpath: &str, content: &[u8]) -> Result<()>;
-
-    /// Get metadata about a file.
+    fn list_dir(&self, relpath: &str) -> Result<ListDir>;
+    fn create_dir(&self, relpath: &str) -> Result<()>;
     fn metadata(&self, relpath: &str) -> Result<Metadata>;
 
     /// Delete a file.
@@ -121,7 +177,13 @@ pub trait Transport: Send + Sync + std::fmt::Debug {
     fn remove_dir_all(&self, relpath: &str) -> Result<()>;
 
     /// Make a new transport addressing a subdirectory.
-    fn sub_transport(&self, relpath: &str) -> Arc<dyn Transport>;
+    fn chdir(&self, relpath: &str) -> Arc<dyn Protocol>;
+
+    fn url(&self) -> &Url;
+
+    fn local_path(&self) -> Option<PathBuf> {
+        None
+    }
 }
 
 /// A directory entry read from a transport.
@@ -173,8 +235,22 @@ pub enum ErrorKind {
     #[display(fmt = "Permission denied")]
     PermissionDenied,
 
+    #[display(fmt = "Unsupported URL scheme")]
+    UrlScheme,
+
     #[display(fmt = "Other transport error")]
     Other,
+}
+
+impl From<io::ErrorKind> for ErrorKind {
+    fn from(kind: io::ErrorKind) -> Self {
+        match kind {
+            io::ErrorKind::NotFound => ErrorKind::NotFound,
+            io::ErrorKind::AlreadyExists => ErrorKind::AlreadyExists,
+            io::ErrorKind::PermissionDenied => ErrorKind::PermissionDenied,
+            _ => ErrorKind::Other,
+        }
+    }
 }
 
 impl Error {
@@ -183,16 +259,10 @@ impl Error {
     }
 
     pub(self) fn io_error(path: &Path, source: io::Error) -> Error {
-        let kind = match source.kind() {
-            io::ErrorKind::NotFound => ErrorKind::NotFound,
-            io::ErrorKind::AlreadyExists => ErrorKind::AlreadyExists,
-            io::ErrorKind::PermissionDenied => ErrorKind::PermissionDenied,
-            _ => ErrorKind::Other,
-        };
         Error {
+            kind: source.kind().into(),
             source: Some(Box::new(source)),
             path: Some(path.to_string_lossy().to_string()),
-            kind,
         }
     }
 
@@ -227,3 +297,16 @@ impl error::Error for Error {
 }
 
 type Result<T> = result::Result<T, Error>;
+
+#[cfg(test)]
+mod test {
+    use std::path::Path;
+
+    use super::Transport;
+
+    #[test]
+    fn get_path_from_local_transport() {
+        let transport = Transport::local(Path::new("/tmp"));
+        assert_eq!(transport.local_path().as_deref(), Some(Path::new("/tmp")));
+    }
+}